requirements.txt
alphabet-detector==0.0.7
bleach==4.1.0
cachetools==4.2.2
CurrencyConverter
emoji==1.6.1
feedparser==6.0.8
pyrate-limiter==2.4.7
lxml==4.6.4
nltk==3.6.5
pillow>=4.2.0
pretty-errors==1.2.24
psutil==5.8.0
pynewtonmath==1.0.1
python-telegram-bot==13.11
regex==2021.11.2
requests==2.26.0
spamwatch==0.3.0
speedtest-cli==2.1.3
spongemock==0.3.4
sqlalchemy==1.4.26
telethon==1.23.0
ujson==4.2.0
wikipedia==1.4.0
zalgo-text==0.6
fontTools==4.27.1
nekos.py==1.0.3
telegraph==2.0.1
tgcrypto==1.2.2
youtube_search==2.1.0
pytube==11.0.1
pymongo==3.12.1
motor>=2.1.0,<2.6.0
dnspython==2.1.0
redis==3.5.3
gtts==2.2.3
cloudscraper==1.2.58
wget==3.2
tswift==0.7.0
lyricsgenius==3.0.1
envparse==0.2.0
cryptg==0.2.post4
aioredis==2.0.0
odmantic==0.3.5
pykeyboard==0.1.2
aiofiles==0.7.0
hachoir==3.1.2
bing_image_downloader==1.1.1
pytz==2021.3
html2text==2020.1.16
NumPy==1.21.3
opencv-python-headless==4.5.4.58
uvloop==0.16.0
python-dotenv==0.19.1
gpytranslate==1.3.0
glitch_this==1.0.2
countryinfo==0.1.2
faker==9.7.1
PyDictionary==2.0.1
ffmpeg==1.4
python-rapidjson==1.5
markdown2>=2.4.0
cowpy==1.1.1
selenium==4.0.0
hiredis==2.0.0
pendulum==2.1.2
img2pdf==0.4.3
google-api-python-client==2.29.0
secureme==0.0.4
git+https://github.com/TeamScenario/GoGoAnimeAPI.git
better_profanity==0.7.0
pydub==0.25.1
nudepy==0.5.1
textblob==0.17.1
requests_html==0.10.0
bs4==0.0.1
multicolorcaptcha==1.1.3
launchpadlib==1.10.15.1
google-trans-new==1.1.9
jikanpy==4.3.2
futures==3.1.1
future
psycopg2-binary
geopy
humanize
heroku3
beautifulsoup4
pyrogram==1.4.16
yt-dlp
python-arq
aiohttp
carbonnow
search_engine_parser
googletrans
youtube_search_python
httpx
gitpython
loguru
